pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "jarvas"
version = "0.5.0"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.0",
    "python-dotenv>=1.0",
    "rich>=13.0",
    "prompt_toolkit>=3.0",
    "google-generativeai>=0.8",
    "supabase>=2.0",
    "requests>=2.31",
    "httpx>=0.27",
    "aiofiles>=23.0",
    "uvicorn[standard]>=0.29",
]

[project.scripts]
jarvas = "jarvas.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.setuptools.packages.find]
where = ["."]
include = ["jarvas*"]
